requirements.txt
# This file may be used to create an environment using:
# $ conda create --name gcn --file requirements.txt --channel conda-forge --channel bioconda --channel pytorch --channel pyg
jupyterlab
scikit-learn
numpy
pandas
matplotlib
scipy
pytorch
torchvision
msprime
stdpopsim
tskit
pytorch-lightning
pyg
pytorch-scatter
pytorch-sparse
networkx
ipywidgets
gpustat
torchvision
torchview
